add my ubuntu-1 kvm to my ssh config
only works from higgins, of course. may switch it to be bridged
later, but for now, only from higgins.

this machine is sort of a playground. sort of. I would like to make
it so it's a test box for setting up a shell provider, lock it down
really good and whatnot. getting there.