dear vim: don't unindent my comments
vim with smartindent has a really annoying habit of removing
any indentation I had inherited from above when I insert a comment
(or, at least #) and then I have to fix the indentation

this fixes it. sorta. it's really hacky, and I may just shut smartindent
off, especially since it doesn't seem to be honoring my backspace backs up
one level of indentation thing, but we shall see.

one part of a long string of changes to my vimrc to make it sane for me