set up indentation settings for ruby/php/puppet
I need to find a better way to do this that's not so global, yet doesn't require
adding a vim modeline to every single file I touch.

I don't like that this is a global setting because if I work on an
outside project and they have their own indentation standards, I don't want
to have to fuck around with all of those. Sadly, the handful of plugins
I tried which would pull in vim settings that we could check into svn into
our projects either completely did not work (as far as I could tell at least)
or were very limited and/or I couldn't figure out how to use them.

I will continue on my quest, however.