separating mutt into different rc files for accounts
I don't feel like trying to manage multiple imap accounts in a single
mutt instance, but I would like to have a lot of things be shared
between them. This is my method for doing that, and it seems to be
working pretty well. I have site-specific configs and shared configs and
the site-specific config includes the shared config. seems to be working
pretty nicely so far