enable line numbers only on prog-mode and conf-mode
I did a lisp!

This one took some doing as I originally was trying to use
https://www.emacswiki.org/emacs/LineNumbers#toc1 to get me the line
numbers I wanted. But the problem with that is, doom turns off all of
the automatic line number mode hook things and enables its own, which I
only found out after digging through the doom code. That's fine, though.
I managed to undo *their* hooks and install my own :grin:

I may open a PR to them proposing that at the very least the list of
hooks is configurable as it's currently hard coded and seemingly the
only way to do what I did was to ... do what I did, which is really
ugly, imo, but it works! Yay!