re-enable clipetty, it "works" with tmux
It doesn't actually do proper clipboardy things in tmux, but at least it
doesn't error and/or block operations that would trigger clipetty

wrapping with tmux is a reasonable thing to do, so that's fine. I will
check with blinksh about the possibility of supporting those clipboard
operations in the future (maybe I'll even try to add the support myself?
*grin*)