changed status line a bit
tbh, the off-center window names were driving me batty :)

They get centered between the status-right and status-left text, and
since the status-left was much longer than the status-right, they were
offset to the right a bit. But having the hostname on the right and the
session name on the left means it's more balanced. Yay!