take `--hidden` out of ripgrep
I must say, I'm not actually the biggest fan of ripgrep. I feel like
silversearcher was more intuitive?

Like I want it to search "hidden" files, but not like ... VCS repo files
and such (with `--hidden` here, it shows results in `.git`, etc.) but I
really don't want to have to re-add ignores for all of the things it
would normally ignore if not for `--hidden` .. that just seems ... I'll
have to do some more reading.