friendlier pohara and conf-space-mode
When I `ssh pohara` I want to go to my rpi. 99% of the time, I'm doing
that at home, and I want the local IP. I'm actually totally fine with
connecting to my external IP/port and forwarding through, I don't really
care. The problem is, my DNS for pohara.kitchen.io is *always* wrong.
And I haven't bothered to fix that. So instead I've just been constantly
sshing to the IP, which is really annoying to type. So I'll make it so
`pohara` defaults to the internal IP and call it good.

Also, make it so files in .ssh/config.d/ are auto-mode as ssh
config (which appears to be conf-space-mode) for proper indentation and suchlike