couple fixes to signmebro
previously, no matter the input file, the signed certificate was getting
written out to the ed25519-cert.pub file. oops. This better supports
non-ed25519 files. I want to continue this and have it have a search
pattern or something that it uses so I can support other certificate
types in the future (e.g. ecdsa-sk, since my yubikey neo doesn't support
ed25519-sk, it seems) but for now this is fine. The only device I ssh
into for non-work things at the moment doesn't support -sk options, so
it's moot :)

Also, I made `vault login` stop dumping my token to the terminal when I
use it. This should make it safer to make a demo screencast or something
if I want to.