add some zmk extensions to be dts (devicetree) files
I would love to figure out if there's a way to do this on a per project
basis similar to how helix did it and similar to editorconfig.

I know editorconfig is a thing but it doesn't have the capability of
setting an editor filetype as far as I can tell.

I don't like having these be globally overridden, in case there are
conflicts between projects. Also I'm not entirely certain this is the
correct way to do this without stomping all over the existing devicetree
definition (with helix I had to basically copy the entire upstream
definition over just to add some file extensions, oof)